now the fps slider is an int

+++ Vroom/Application.h
@@ -36,7 +36,7 @@ public:
 	update_status Update();
 	bool CleanUp();
 	bool RequestBrowser(const char* path);
-	float maxFps;
+	int maxFps;
 private:
 
 	void AddModule(Module* mod);
